Module elvis::core

Foundational abstractions for building Internet simulations.

This module contains the necessary pieces to implement protocols and to simulate machines communicating across networks. Elvis follows the x-kernel design for protocol layering.

Organization

Message and Control provide basic utilities common to most protocols

Protocol and Session implement individual protocols

Internet provides the actual simulation

Protocol structure

Protocol and Session work closely together. A session contains the state for a single open connection on a single protocol. For example, a TCP session would contain information about the window, the state of the connection, and the stream of bytes to send. Sessions are created by the protocol either in response to a program opening a connection or a new connection being opened for a listening server program. In addition to creating new sessions, protocols also route incoming packets to the correct sessions.

Re-exports

pub use control::Control;

pub use message::Message;

Modules

control

Types for exchanging data between protocols.

message

Byte collections with efficient operations for protocols.

Structs

Internet

The top-level container that controls the simulation.

ProtocolContext

Provides a Protocol with information about its execution environment.

ProtocolId

A unique identifier for a Protocol.

SharedSession

A shared handle to a Session.

Traits

Protocol

A member of a networking protocol stack.

Session

Holds the state for a particular connection.

Type Definitions

SharedProtocol

A shared handle to a Protocol.