Trait elvis::core::Protocol

pub trait Protocol {
    fn id(&self) -> ProtocolId;
    fn open(
        &mut self, 
        upstream: ProtocolId, 
        participants: Control, 
        context: &mut ProtocolContext
    ) -> Result<SharedSession, Box<dyn Error>>;
    fn listen(
        &mut self, 
        upstream: ProtocolId, 
        participants: Control, 
        context: &mut ProtocolContext
    ) -> Result<(), Box<dyn Error>>;
    fn demux(
        &mut self, 
        message: Message, 
        context: &mut ProtocolContext
    ) -> Result<(), Box<dyn Error>>;

    fn start(
        &mut self, 
        _context: ProtocolContext, 
        _shutdown: Sender<()>
    ) -> Result<(), Box<dyn Error>> { ... }
}

A member of a networking protocol stack.

A protocol is responsible for creating new Sessions and demultiplexing requests to the correct session.

Required Methods

fn id(&self) -> ProtocolId

Returns a unique identifier for the protocol.

fn open(
&mut self,
upstream: ProtocolId,
participants: Control,
context: &mut ProtocolContext
) -> Result<SharedSession, Box<dyn Error>>

Actively open a new network connection.

Called by the upstream protocol to create a new Session for a connection. Each protocol should, in turn, open a session with some downstream protocol to establish a chain of sessions with which to send and receive messages for the requesting user program.

The participants set contains key-value pairs that identify aspects of a connection to facilitate demuxing. It should contain all attributes needed to uniquely identify the connection. For example, an IP protocol might require the attributes {local_address, remote_address}. A UDP or TCP protocol might require the attributes {local_address, local_port, remote_address, remote_port}.

fn listen(
&mut self,
upstream: ProtocolId,
participants: Control,
context: &mut ProtocolContext
) -> Result<(), Box<dyn Error>>

Listen for new connections.

Requests that messages for which there is no existing session be sent to the upstream protocol. Only messages that match the participants set will be forwarded. See demux for more details.

The participants set should contain all attributes needed to identify the listening program. For example, an IP protocol might use the set of attributes {local_address}. Since we are listening for connections from any remote address, when the IP protocol sees a message it does not have a session for, it will check whether the local address given in the header is one it is listening for. If so, it will create the session identified by {local_address, remote_address} and continue demultiplexing the message. Similarly, a UDP or TCP protocol would want its participant set to include {local_address, local_port}.

fn demux(
&mut self,
message: Message,
context: &mut ProtocolContext
) -> Result<(), Box<dyn Error>>

Identifies the session that a message belongs to and forwards the message to it.

When demultiplexing a message, a protocol will typically carry out several tasks:

Remove and parse the message header.

Apply information about the message header to the context. This should include any information that the target session or other protocols may need to know about. For example, an IP protocol should add the source and destination addresses to the context so that UDP and TCP may use them for verifying checksums.

Select a session to respond to the message. This is done by looking at information extracted from the header. If there is no matching session, the protocol should check to see whether any protocol has asked to receive the message by calling listen at an earlier time. If so, a new session should be created.

Call receive on the selected session.

Provided Methods

fn start(
&mut self,
_context: ProtocolContext,
_shutdown: Sender<()>
) -> Result<(), Box<dyn Error>>

Start the protocol running. This may spawn tasks and timers that the protocol needs to run. The default implementation is a no-op.