Trait elvis::protocols::user_process::Application

pub trait Application {
    const ID: ProtocolId;

    fn start(
        &mut self, 
        context: ProtocolContext, 
        shutdown: Sender<()>
    ) -> Result<(), Box<dyn Error>>;
    fn recv(
        &mut self, 
        message: Message, 
        context: &mut ProtocolContext
    ) -> Result<(), Box<dyn Error>>;
}

A program being run in a UserProcess.

An application is similar to a stripped-down Session. It runs when messages come in over the network or when the containing machine awakens the application to give it time to run.

Required Associated Constants

const ID: ProtocolId

A unique identifier for the application.

Required Methods

fn start(
&mut self,
context: ProtocolContext,
shutdown: Sender<()>
) -> Result<(), Box<dyn Error>>

Gives the application time to run. Unlike recv, awake is not called in response to specific events.

fn recv(
&mut self,
message: Message,
context: &mut ProtocolContext
) -> Result<(), Box<dyn Error>>

Called when the containing UserProcess receives a message over the network and gives the application time to handle it.